Otherwise trouble with pythonization in xAH
Former-commit-id: 446ccc387ddbbd6608abb78fdcc5d7cb3120672f

+++ Reconstruction/Jet/JetRecTools/JetRecTools/JetConstituentModSequence.h
@@ -40,7 +40,7 @@ protected:
   // P-A : the actual type
   // Define as a basic integer type because Gaudi
   // doesn't support arbitrary property types
-  unsigned short m_inputType; // 
+  int m_inputType; // 
   
   
   ToolHandleArray<IJetConstituentModifier> m_modifiers;
